+++ test/src/read_write.cpp
@@ -7,19 +7,19 @@
 #include "tree/nodes.hpp"
 
 
-template<uint8_t I, bool B, NodeLike... N>
+template<uint16_t I, bool B, NodeLike... N>
 struct NodeHeaderImpl_{
     static constexpr uint8_t ID = I;
 
     using childrenTypes = std::tuple<N...>;
 
-    bool guard(const uint8_t& queryID){
+    bool guard(const uint16_t& queryID){
       return queryID == ID ? true : false;
     }
 
 };
 
-template<uint8_t I, auto V, class T>
+template<uint16_t I, auto V, class T>
 struct LeafnodeHeaderImpl{
     static constexpr uint8_t ID = I;
